requirements.txt
azure-storage-blob>=12.16.0
azure-ai-formrecognizer>=3.2.1
python-docx>=0.8.11
PyPDF2>=3.0.0
python-dotenv==1.0.0
requests>=2.31.0
google-generativeai>=0.3.0